tests/wpt/tests/css/css-text/parsing/text-autospace-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text: getComputedStyle().textAutospace</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#propdef-text-autospace">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("text-autospace", "normal");
test_computed_value("text-autospace", "no-autospace");
test_computed_value("text-autospace", "auto");

test_computed_value("text-autospace", "ideograph-alpha");
test_computed_value("text-autospace", "ideograph-numeric");

</script>
</body>
</html>
